tests: fix travis 'script' to fail tests correctly
The return status of a loop is that of the last command executed in the loop.
That means previously, the build would not fail during tests unless the last
test script failed. Any/all of the other tests could have failed and the build
would still pass.

This commit does two things:
    1) explicitly execute the test file with `... sh $file ...`
    2) exit (non-zero) the tests loop on a failing test