Make it so that conf/*.sh can be trivially overridden by the user
Closes #90

Prior to this commit, editing the conf/*.conf files resulted in
changed configuration parameters, but editing the conf/*.sh files did
not.  Furthermore, there was a comment in the conf/*.sh files that
could be misconstrued.

With this commit the conf/*.sh and conf/*.conf files behave the same.
conf/*.sh files can be edited and then docker -v can be used to bind
mount the conf directory and the changes will take effect.