tests: fix failing `customConfig` test
The custom configuration check in `conf/config.sh` checks for the presence of a
`conf.d` directory, but is not very specific as to where that directory is to
be expected.  Docker users are expected to mount the `conf.d` directory under
`$INSPIRCD_ROOT` set in the entrypoint script (`/inspircd`), but because the
entrypoint script is running at `/`, the subsequent `exec ...` starts (and
stays in) that directory. Thus, `conf.d` is never found.

The commit passes the `$INSPIRCD_ROOT` to the `exec`'d `inspircd` allowing the
`conf/config.sh` script to (use and) look for `conf.d` in the expected
location. Additionally, the working dir is changed to `$INSPIRCD_ROOT` before
the `exec ...` call, so that any `<include executable="...`s will be correct
path `/inspircd/conf`, as opposed to `/conf` (which is the base config template
dir).